refactor: extract datapath/interface converters
