+++ proto/lib.rs
@@ -1,8 +1,11 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 
 mod codec;
+pub mod convert;
 mod interface;
+mod limits;
 mod protobuf;
+pub mod time;
 
 pub use crate::codec::decode_key;
 pub use crate::codec::encode_key;
